% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_compose.r
\name{add_mutations}
\alias{add_mutations}
\title{Add mutations to circos plot}
\usage{
add_mutations(
  circos,
  mutations,
  col_chrom = "chromosome",
  col_pos = "position",
  col_vaf = "vaf",
  col_type = NULL,
  col_tooltip = NULL
)
}
\arguments{
\item{circos}{a circos object created using \code{ggcircos()}}

\item{col_chrom}{Name of column describing the chromosome of the mutation}

\item{col_pos}{Name of column describing the 1-based position of the mutation.}

\item{col_vaf}{Name of column}

\item{col_type}{Name of column describing the type of mutation. Used to colour mutations. If NULL mutations will not be coloured.}

\item{data}{a data.frame where each row represents a mutations.}
}
\value{
a Circos class object with mutations added
}
\description{
Add mutations to circos plot
}
\examples{
## Filepaths
fai <- system.file("hg38.fai", package = "ggcircos")
mutations_tsv <- system.file("hg38.mutations.fai", package = "ggcircos")

# Read in data
hg38 <- fai_to_reference(fai)
mutations <- read.csv(mutations_tsv, sep = "\t")

# Create Circos Plot
multiomic_circos <- ggcircos(hg38) |>
  add_mutations(mutations)
}
