% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_mapping.r
\name{create_normalisation_map}
\alias{create_normalisation_map}
\title{Create a data structure mapping chr + original positions to a new mega-interval}
\usage{
create_normalisation_map(stdref, inter_chrom_spacing, target_range = c(1, 100))
}
\arguments{
\item{stdref}{data.frame with columns 'name', 'start', 'end' where intervals are 1-based both-end inclusive.}

\item{inter_chrom_spacing}{how much space is between each chromosome}

\item{target_range}{c(1, 100)}
}
\value{
a e.g.
list(
chr1 = list(
original = c(start=start, end=end),
new = c(start=start, end=end),
),
chr2 = list(
original = c(start=start, end=end),
new = c(start=start, end=end),
)
)
}
\description{
Take the original start & end for each chromosome, line them up into a single mega interval (with \code{inter_chrom_spacing}).
Returns a data structure mapping each chromosome to a list of two ranges,
one representing the original range, the other representing the new range.
This makes it easy to rescale a chr + postion using \code{\link[=project_points]{project_points()}}
}
