% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-link-lab.R
\name{geom_link_lab}
\alias{geom_link_lab}
\title{the node label layer of circos}
\usage{
geom_link_lab(data = NULL, mapping = NULL, hjust = 0, offset = 0, ...)
}
\arguments{
\item{data}{input data}

\item{mapping}{mapping aesthetics}

\item{hjust}{horizontal adjustment to nudge labels}

\item{offset}{Horizontal adjustment to nudge labels by.}

\item{...}{additional parameters.}
}
\description{
the node label layer of circos
}
