\name{ggcircos-package}
\alias{ggcircos-package}
\alias{ggcircos}
\docType{package}
\title{
\packageTitle{ggcircos}
}
\description{
\packageDescription{ggcircos}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggcircos}
\packageIndices{ggcircos}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggcircos}

Maintainer: \packageMaintainer{ggcircos}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
