% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_mapping.r
\name{project_intervals}
\alias{project_intervals}
\title{Project intervals onto circos mega-interval}
\usage{
project_intervals(chr, start, end, map, return_type = c("list", "data.frame"))
}
\arguments{
\item{chr}{chromosome name}

\item{start}{start position in original coord system (1 based)}

\item{end}{end position in original coord system (1 based)}

\item{map}{produced by \code{\link[=create_normalisation_map]{create_normalisation_map()}}. A list mapping chromosomes to original & new ranges.}

\item{return_type}{Output format. Either \code{"list"} to return a list of 2-element numeric vectors, or \code{"data.frame"} to return a data frame with columns \code{chr}, \code{start}, and \code{end}.}
}
\value{
If \code{return_type = "list"}, a list of 2-element numeric vectors representing projected start and end positions. If \code{return_type = "data.frame"}, a data frame with columns \code{chr}, \code{start}, and \code{end}.
}
\description{
Note target range is defined by \code{target_range} argument of \code{create_normalisation_map}
}
