% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_mapping.r
\name{project_points}
\alias{project_points}
\title{Project points onto circos mega-interval}
\usage{
project_points(chr, pos, map)
}
\arguments{
\item{chr}{chromosome name}

\item{pos}{position in original coord system (1 based)}

\item{map}{produced by \code{\link[=create_normalisation_map]{create_normalisation_map()}}. A list mapping chromosomes to original & new ranges.}
}
\value{
a single number representing position in unified (mega-interval) coord system.
}
\description{
Note target range is defined by \code{target_range} argument of \code{create_normalisation_map}
}
