% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_mutations}
\alias{simulate_mutations}
\title{Simulate mutations based on a reference}
\usage{
simulate_mutations(
  reference,
  col_name = "chromosome",
  col_length = "length",
  nmuts = 100,
  types = c("C>A", "C>G", "C>T", "T>A", "T>C", "T>G"),
  type_probability = rep(1/length(types), times = length(types))
)
}
\arguments{
\item{reference}{a 2-column data.frame describing the name and length of each reference sequence.
Can create from .fai index file using \code{\link[=fai_to_reference]{fai_to_reference()}}}

\item{col_name}{name of column in reference data.frame that describes sequence name}

\item{col_length}{name of column in reference data.frame that describes sequence length}

\item{nmuts}{number of mutations to simulate}

\item{types}{a vector of possible mutation 'types'}

\item{type_probabilty}{a vector describing the probabiliy of assigning a mutation to each type.
Should be the same length as types vector.}
}
\value{
A mutations data.frame compatible with the \code{\link[=add_mutations]{add_mutations()}} function. Contains the columns:
\enumerate{
\item chromosome
\item position
\item vaf
\item type
}
}
\description{
Simulates single base mutations based on a reference genome.
}
\details{
Simulation is based on uniform distribution of mutations across the full length of the geome.
VAFs are normally distributed around 0.5
Type column is based on random sampling aginst
}
\examples{
## Get Reference Based on fai index
fai <- system.file("hg38.fai", package = "ggcircos")
hg38 <- fai_to_reference(fai)

# Simulate 100 mutations
mutations <- simulate_mutations(hg38, nmuts = 100)


Plot simulated mutations
multiomic_circos <- ggcircos(hg38) |>
  add_mutations(mutations)
}
