% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_compose.r
\name{standardise_reference_regions}
\alias{standardise_reference_regions}
\title{Standardise reference regions}
\usage{
standardise_reference_regions(
  reference,
  col_name = "chromosome",
  col_length = "length"
)
}
\arguments{
\item{reference}{a 2-column data.frame describing the name and length of each reference sequence.
Can create from .fai index file using \code{\link[=fai_to_reference]{fai_to_reference()}}}

\item{col_name}{name of column in reference data.frame that describes sequence name}

\item{col_length}{name of column in reference data.frame that describes sequence length}
}
\value{
4-column data.frame with columns 'name, start, end, refindex'.
Intervals are 1-based & both-end inclusive ([start, end]).
This means a [1, 1000] reference entry is a 1000 bp-long region.
refindex is a unique number from 1 to nrow that is unique for each reference chromosome name and is used
by downstream functions to convert chrom-pos positions into a single coord position on x-axis. Based on order in input reference data.frame.
}
\description{
Standardise reference regions
}
\examples{
df = data.frame(chromosome = c("chr1", "chr2"), length = c(1000, 2000))
reference <- standardise_reference_regions(df)
print(reference)

}
