% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_compose.r
\name{transform_position_to_reference}
\alias{transform_position_to_reference}
\title{Create Coordinate Transformer}
\usage{
transform_position_to_reference(reference, inter_chrom_gap = 0)
}
\arguments{
\item{reference}{standardised reference data.frame created from \code{\link[=standardise_reference_regions]{standardise_reference_regions()}}}

\item{inter_chrom_gap}{proportion of total genome size that will be taken up by gaps between chromosomes}
}
\value{
A function that takes two arguments 'chromosome' and 'position' and returns a single number representing position on the new x axis
}
\description{
Once we have a reference karyotype set, we to transform chrom + pos genomic positions to a single x-coordinate position
that tells us where on the x axis to plot.
For example a variant at chr1:100 we might want to plot at position 100 on the x axis, whereas chr2:50 we might want to plot at
x = chr1_length + 50.
If we want our x coord axis to have gaps between each chromosome, we might need to take into account additional buffer space.
}
