% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_compose.r
\name{fai_to_reference}
\alias{fai_to_reference}
\title{Parse fai index to reference data.frame}
\usage{
fai_to_reference(fai)
}
\arguments{
\item{fai}{Path to a .fai index file created by running \verb{samtools faidx <reference.fasta>}.}
}
\value{
A 2-column data.frame with fields 'chromosome' and 'length' that can be parsed by ggcircos
}
\description{
Parse fai index to reference data.frame expected by ggcircos
}
\examples{
fai <- system.file("hg38.fai", package = "ggcircos")
fai_to_reference(fai)
}
