% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_compose.r
\name{ggcircos}
\alias{ggcircos}
\title{Create a base circos plot}
\usage{
ggcircos(
  reference,
  col_name = "chromosome",
  col_length = "length",
  inter_chrom_spacing = 0.1
)
}
\arguments{
\item{reference}{a 2-column data.frame describing the name and length of each reference sequence.
Can create from .fai index file using \code{\link[=fai_to_reference]{fai_to_reference()}}}

\item{col_name}{name of column in reference data.frame that describes sequence name}

\item{col_length}{name of column in reference data.frame that describes sequence length}
}
\value{
A ggplot canvas with karyotype information
}
\description{
Create a base circos plot
}
\examples{
## Filepaths
fai <- system.file("hg38.fai", package = "ggcircos")
mutations_tsv <- system.file("hg38.mutations.fai", package = "ggcircos")

# Read in data
hg38 <- fai_to_reference(fai)
mutations <- read.csv(mutations_tsv, sep = "\t")

# Create Circos Plot
multiomic_circos <- ggcircos(hg38) |>
  add_mutations(mutations)
}
