% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcircos.R
\name{ggcircos}
\alias{ggcircos}
\title{visualizing relationship data and heterogenous associated data based on grammar of graphics}
\usage{
ggcircos(
  x,
  mapping = NULL,
  layout = "circular",
  radius = 1,
  ncp = 4,
  hratio = 0.5,
  alpha = 0.4,
  ...
)
}
\arguments{
\item{x}{data}

\item{mapping}{aesthetic mapping}

\item{layout}{character, 'linear' or 'circular'}

\item{radius}{numeric the radius of the circular layout}

\item{ncp}{numeric}

\item{hratio}{numeric}

\item{alpha}{numeric}

\item{...}{additional parameter}
}
\description{
visualizing relationship data and heterogenous associated data based on grammar of graphics
}
\examples{
data(mtcars)
mtcars \%>\% cor \%>\% as.tbl_lk() \%>\% ggcircos()
}
